put the Avatar component in Storybook
